Rename the `cli` package to `params`
