+++ cli/src/main/kotlin/io/spine/protodata/cli/app/Main.kt
@@ -44,16 +44,16 @@ import io.spine.logging.WithLogging
 import io.spine.logging.context.LogLevelMap
 import io.spine.logging.context.ScopedLoggingContext
 import io.spine.protodata.backend.Pipeline
-import io.spine.protodata.cli.DebugLoggingParam
-import io.spine.protodata.cli.InfoLoggingParam
-import io.spine.protodata.cli.Parameter
-import io.spine.protodata.cli.PluginParam
-import io.spine.protodata.cli.ProtoFilesParam
-import io.spine.protodata.cli.RequestParam
-import io.spine.protodata.cli.SettingsDirParam
-import io.spine.protodata.cli.SourceRootParam
-import io.spine.protodata.cli.TargetRootParam
-import io.spine.protodata.cli.UserClasspathParam
+import io.spine.protodata.params.DebugLoggingParam
+import io.spine.protodata.params.InfoLoggingParam
+import io.spine.protodata.params.Parameter
+import io.spine.protodata.params.PluginParam
+import io.spine.protodata.params.ProtoFilesParam
+import io.spine.protodata.params.RequestParam
+import io.spine.protodata.params.SettingsDirParam
+import io.spine.protodata.params.SourceRootParam
+import io.spine.protodata.params.TargetRootParam
+import io.spine.protodata.params.UserClasspathParam
 import io.spine.protodata.plugin.Plugin
 import io.spine.protodata.protobuf.ProtoFileList
 import io.spine.protodata.render.SourceFileSet
