+++ params/src/main/kotlin/io/spine/protodata/params/Parameter.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.protodata.cli
+package io.spine.protodata.params
 
 import java.io.File.pathSeparator
