+++ params/src/main/proto/spine/protodata/params/pipeline_parameters.proto
@@ -26,7 +26,7 @@
 
 syntax = "proto3";
 
-package spine.protodata;
+package spine.protodata.params;
 
 import "spine/options.proto";
